requirements.txt
streamlit
PyPDF2
pdf2image
pytesseract
pillow
opencv-python
python-docx
google-cloud-vision
google-generativeai
